Добавление метода GetCity к таблице City

+++ TravelApp.API/Controllers/CityController.cs
@@ -50,12 +50,26 @@ namespace TravelApp.API.Controllers
         }
         [HttpDelete("{id}")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public async Task<ActionResult<City>> Delete(int id)
+        public async Task<ActionResult> Delete(int id)
         {
             if (id <= 0) return BadRequest();
+            if (await _repository.GetAsync(id) == null) return NotFound();
             await _repository.DeleteCityAsync(id);
             return NoContent();
+        }
+
+        [HttpGet("{id}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public async Task<ActionResult<City>> Get(int id)
+        {
+            if (id <= 0) return BadRequest();
+
+            var result = await _repository.GetAsync(id);
+            return result == null ? NotFound() : Ok(result);
         }
